requirements.txt
streamlit
langchain
openai
python-dotenv
pydantic